# spec/api_contracts.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Virtual Try-On API",
    "description": "Comprehensive API specification for the Virtual Try-On application supporting avatar scanning, 3D product try-on, e-commerce, and user management",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "api-support@tryon.com",
      "url": "https://docs.tryon.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.tryon.com/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.tryon.com/v1",
      "description": "Staging server"
    },
    {
      "url": "https://dev-api.tryon.com/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/scan": {
      "post": {
        "tags": ["scan"],
        "summary": "Start avatar scanning",
        "description": "Initiate 3D avatar scanning process using device camera",
        "operationId": "startAvatarScan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              },
              "examples": {
                "photo_scan": {
                  "value": {
                    "scanType": "photo_based",
                    "images": [
                      {
                        "url": "https://storage.tryon.com/scans/tmp_1.jpg",
                        "pose": "front",
                        "quality": 0.95
                      },
                      {
                        "url": "https://storage.tryon.com/scans/tmp_2.jpg",
                        "pose": "side",
                        "quality": 0.92
                      },
                      {
                        "url": "https://storage.tryon.com/scans/tmp_3.jpg",
                        "pose": "back",
                        "quality": 0.88
                      }
                    ],
                    "userPreferences": {
                      "bodyType": "athletic",
                      "height": 175,
                      "weight": 70,
                      "preferences": ["casual", "business"]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Scan initiated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "scanId": "scan_789xyz123",
                        "status": "processing",
                        "estimatedTimeMinutes": 5,
                        "queuePosition": 3,
                        "createdAt": "2025-01-15T10:30:00Z",
                        "webhookUrl": "https://app.tryon.com/webhooks/scan-complete"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid scan request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded - too many scan requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products": {
      "get": {
        "tags": ["products"],
        "summary": "Get products catalog",
        "description": "Retrieve paginated list of products with filtering and sorting",
        "operationId": "getProducts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "example": 20
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by product category",
            "schema": {
              "type": "string",
              "enum": ["tops", "bottoms", "dresses", "outerwear", "shoes", "accessories", "activewear", "formal"]
            },
            "example": "tops"
          },
          {
            "name": "subcategory",
            "in": "query",
            "description": "Filter by subcategory",
            "schema": {
              "type": "string",
              "examples": ["shirts", "pants", "jackets", "sneakers", "hats"]
            }
          },
          {
            "name": "brand",
            "in": "query",
            "description": "Filter by brand name",
            "schema": {
              "type": "string"
            },
            "example": "Nike"
          },
          {
            "name": "priceMin",
            "in": "query",
            "description": "Minimum price filter",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "example": 29.99
          },
          {
            "name": "priceMax",
            "in": "query",
            "description": "Maximum price filter",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "example": 199.99
          },
          {
            "name": "size",
            "in": "query",
            "description": "Filter by available sizes",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["XS", "S", "M", "L", "XL", "XXL", "6", "7", "8", "9", "10", "11", "12"]
              }
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "color",
            "in": "query",
            "description": "Filter by available colors",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "availability",
            "in": "query",
            "description": "Filter by product availability",
            "schema": {
              "type": "string",
              "enum": ["in_stock", "out_of_stock", "pre_order"],
              "default": "in_stock"
            },
            "example": "in_stock"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort results by field",
            "schema": {
              "type": "string",
              "enum": ["name", "price_asc", "price_desc", "rating", "newest", "popularity"],
              "default": "popularity"
            },
            "example": "popularity"
          },
          {
            "name": "tryonAvailable",
            "in": "query",
            "description": "Only show products available for virtual try-on",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "example": true
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "products": [
                          {
                            "id": "prod_123abc456",
                            "name": "Premium Cotton T-Shirt",
                            "description": "High-quality cotton blend t-shirt with modern fit",
                            "price": 49.99,
                            "originalPrice": 69.99,
                            "currency": "USD",
                            "brand": "FashionBrand",
                            "category": "tops",
                            "subcategory": "shirts",
                            "images": [
                              {
                                "url": "https://cdn.tryon.com/products/prod_123abc456_front.jpg",
                                "type": "front",
                                "resolution": "high"
                              },
                              {
                                "url": "https://cdn.tryon.com/products/prod_123abc456_back.jpg",
                                "type": "back",
                                "resolution": "high"
                              }
                            ],
                            "model3D": {
                              "url": "https://cdn.tryon.com/models/prod_123abc456.glb",
                              "format": "glb",
                              "size": "8.2MB",
                              "polycount": 8450
                            },
                            "variants": [
                              {
                                "id": "var_123_001",
                                "size": "M",
                                "color": "blue",
                                "colorHex": "#2563EB",
                                "stock": 25,
                                "sku": "TSH-001-M-BLU"
                              },
                              {
                                "id": "var_123_002",
                                "size": "L",
                                "color": "blue",
                                "colorHex": "#2563EB",
                                "stock": 18,
                                "sku": "TSH-001-L-BLU"
                              }
                            ],
                            "rating": {
                              "average": 4.3,
                              "count": 127
                            },
                            "tryonAvailable": true,
                            "createdAt": "2025-01-10T12:00:00Z",
                            "updatedAt": "2025-01-14T16:30:00Z"
                          }
                        ],
                        "pagination": {
                          "currentPage": 1,
                          "totalPages": 25,
                          "totalItems": 487,
                          "itemsPerPage": 20,
                          "hasNext": true,
                          "hasPrevious": false
                        },
                        "facets": {
                          "categories": {
                            "tops": 156,
                            "bottoms": 98,
                            "dresses": 67,
                            "outerwear": 45
                          },
                          "brands": {
                            "FashionBrand": 123,
                            "TrendStyle": 89,
                            "ClassicWear": 67
                          },
                          "priceRange": {
                            "min": 19.99,
                            "max": 299.99
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cart": {
      "get": {
        "tags": ["cart"],
        "summary": "Get shopping cart",
        "description": "Retrieve the user's current shopping cart with all items",
        "operationId": "getCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cart retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["cart"],
        "summary": "Clear cart",
        "description": "Remove all items from the shopping cart",
        "operationId": "clearCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cart cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/closet": {
      "get": {
        "tags": ["closet"],
        "summary": "Get user closet",
        "description": "Retrieve the user's virtual closet with owned and liked items",
        "operationId": "getUserCloset",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter closet items by type",
            "schema": {
              "type": "string",
              "enum": ["owned", "liked", "wishlist", "all"],
              "default": "all"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by product category",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Closet retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from login endpoint"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      }
    },
    "schemas": {
...
      "ScanRequest": {
        "type": "object",
        "required": ["scanType", "images"],
        "properties": {
          "scanType": {
            "type": "string",
            "enum": ["photo_based", "video_based", "body_measurement"],
            "description": "Type of scan to perform"
          },
          "images": {
            "type": "array",
            "minItems": 3,
            "maxItems": 10,
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "format": "uri",
                  "description": "URL to uploaded image"
                },
                "pose": {
                  "type": "string",
                  "enum": ["front", "side_left", "side_right", "back", "custom"],
                  "description": "Body pose in the image"
                },
                "quality": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 1,
                  "description": "Image quality score (0-1)"
                }
              }
            }
          },
          "userPreferences": {
            "type": "object",
            "properties": {
              "bodyType": {
                "type": "string",
                "enum": ["slim", "athletic", "curvy", "plus_size"],
                "description": "User's body type"
              },
              "height": {
                "type": "integer",
                "minimum": 100,
                "maximum": 250,
                "description": "Height in centimeters"
              },
              "weight": {
                "type": "number",
                "minimum": 30,
                "maximum": 300,
                "description": "Weight in kilograms"
              },
              "preferences": {
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": ["casual", "formal", "business", "sport", "party", "sleepwear"]
                }
              }
            }
          }
        }
      },
      "ScanResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "scanId": {
                "type": "string",
                "description": "Unique scan identifier"
              },
              "status": {
                "type": "string",
                "enum": ["queued", "processing", "completed", "failed"],
                "description": "Current scan status"
              },
              "estimatedTimeMinutes": {
                "type": "integer",
                "description": "Estimated processing time in minutes"
              },
              "queuePosition": {
                "type": "integer",
                "description": "Position in processing queue"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "webhookUrl": {
                "type": "string",
                "format": "uri",
                "description": "Webhook URL for scan completion notification"
              }
            }
          }
        }
      },
...
      "ProductListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "products": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/Pagination"
              },
              "facets": {
                "type": "object",
                "properties": {
                  "categories": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "integer"
                    }
                  },
                  "brands": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "integer"
                    }
                  },
                  "priceRange": {
                    "type": "object",
                    "properties": {
                      "min": {
                        "type": "number"
                      },
                      "max": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "CartResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Cart"
          }
        }
      },
...
      "ClosetResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ClosetItem"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/Pagination"
              },
              "stats": {
                "type": "object",
                "properties": {
                  "totalItems": {
                    "type": "integer"
                  },
                  "ownedItems": {
                    "type": "integer"
                  },
                  "likedItems": {
                    "type": "integer"
                  },
                  "wishlistItems": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code for client handling"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "description": "Additional error details"
              }
            }
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": ["VALIDATION_ERROR"]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "properties": {
                  "fields": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "field": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "RateLimitError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": ["RATE_LIMIT_EXCEEDED"]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "integer"
                  },
                  "remaining": {
                    "type": "integer"
                  },
                  "resetTime": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [true]
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Success message"
              }
            }
          }
        }
      },
...
      "Cart": {
        "type": "object",
        "required": ["items", "total", "itemCount"],
        "properties": {
          "id": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "subtotal": {
            "type": "number",
            "minimum": 0
          },
          "tax": {
            "type": "number",
            "minimum": 0
          },
          "shipping": {
            "type": "number",
            "minimum": 0
          },
          "discount": {
            "type": "number",
            "minimum": 0
          },
          "total": {
            "type": "number",
            "minimum": 0
          },
          "itemCount": {
            "type": "integer",
            "minimum": 0
          },
          "couponCode": {
            "type": "string"
          },
          "estimatedDelivery": {
            "type": "string",
            "format": "date"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "totalItems": {
            "type": "integer"
          },
          "itemsPerPage": {
            "type": "integer"
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          }
        }
      },
      "ClosetItem": {
        "type": "object",
        "required": ["id", "productId", "type"],
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "type": {
            "type": "string",
            "enum": ["owned", "liked", "wishlist"]
          },
          "purchaseInfo": {
            "type": "object",
            "properties": {
              "orderId": {
                "type": "string"
              },
              "purchaseDate": {
                "type": "string",
                "format": "date"
              },
              "price": {
                "type": "number"
              },
              "warrantyExpiry": {
                "type": "string",
                "format": "date"
              }
            }
          },
          "notes": {
            "type": "string"
          },
          "usageCount": {
            "type": "integer"
          },
          "lastUsed": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}